refactor: mudar pasta services dentro de WordPress para setup, assim faz mais sentido.
